noclip: refactor
